Fix some GWT build issues
Different SpriteBatch's for the different screens caused a 'no buffer
allocated' exception when transitioning from MenuScreen to
GameplayScreen. Using the same SpriteBatch for both screens fixes the
problem.

Pool.Poolable interface was causing reflection problems when it was
declared at the superclass level in Effect. Moved implementation to
subclasses and add extend-configuration-property tag to gwt.xml files to
resolve the issue.